backend/customers/urls.py
from django.urls import path
from rest_framework_simplejwt.views import TokenRefreshView
from .views import RegisterView, MeView, CustomTokenObtainPairView, RequestOTPView, VerifyOTPView

urlpatterns = [
    path("register/",      RegisterView.as_view(),              name="customer-register"),
    path("login/",         CustomTokenObtainPairView.as_view(), name="customer-login"),
    path("token/refresh/", TokenRefreshView.as_view(),           name="token-refresh"),
    path("me/",            MeView.as_view(),                     name="customer-me"),
    path("otp/request/",   RequestOTPView.as_view(),             name="otp-request"),
    path("otp/verify/",    VerifyOTPView.as_view(),              name="otp-verify"),
]
