backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
djangorestframework-simplejwt>=5.3,<6.0
psycopg2-binary>=2.9,<3.0
Pillow>=10.0,<12.0
python-decouple>=3.8,<4.0
django-filter>=24.0,<25.0
drf-spectacular>=0.27,<1.0
django-redis>=5.4,<6.0
python-jose[cryptography]>=3.3,<4.0
gunicorn>=22.0,<23.0
requests>=2.31,<3.0
sendgrid>=6.11,<7.0
